backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
redis[hiredis]==5.2.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
